archive/archive_python_flask_mvp_2025-06-16/investors/webapp/backend/requirements.txt
Flask
Flask-Cors
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
